refactor: extract Core lifecycle functions into minarch_core.c
Moves Core_getName, Core_open, Core_init, Core_applyCheats,
Core_updateAVInfo, Core_load, Core_reset, Core_unload, Core_quit,
and Core_close out of minarch.c.

Promotes environment_callback to non-static so Core_open can register
it with the loaded core library. Core_load retains its dependency on
SRAM_read/RTC_read (via minarch_saves.h) and Cheats_load/Cheats_free
(via minarch_cheats.h).

minarch.c is now ~4970 lines, down from the original 9278.
